agent/backend/requirements.txt
flask==1.0.2
requests==2.21.0
pyserial==3.4
selenium==3.141.0
flask-cors==3.0.7
pywin32
flask-socketio
